api/requirements.txt
python-dotenv
flask
flask_cors
google.generativeai